3.5.2. Rest API 3. PostController fix

+++ src/main/java/ru/job4j/socialmediaapi/repository/picture/PictureRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.socialmediaapi.repository.picture;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.job4j.socialmediaapi.model.Picture;
 
-public interface PictureRepository extends JpaRepository<Picture, Integer> {
+public interface PictureRepository extends JpaRepository<Picture, Long> {
 }
